pyproject.toml
[tool.poetry]
name = "ui tests"
version = "0.1.0"
description = ""
authors = ["rinat akhmetzyanov"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
webdriver-manager = "^3.4.2"
allure-pytest = "^2.9.43"
Faker = "^8.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
